requirements.txt
asgiref==3.7.2
Django==3.2.23
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1